// src/models/IoK8sApiResourceV1alpha2ResourceClaimSpec/types.ts
import { IoK8sApiResourceV1alpha2ResourceClaimParametersReferenceModelOptions as IoK8sApiResourceV1alpha2ResourceClaimParametersReference } from '../IoK8sApiResourceV1alpha2ResourceClaimParametersReference';

export interface IoK8sApiResourceV1alpha2ResourceClaimSpecModelOptions {
  allocationMode?: string;
  parametersRef?: IoK8sApiResourceV1alpha2ResourceClaimParametersReference;
  resourceClassName: string;
}